wwan: simplify API of nm_modem_act_stage2_config() to never fail
The previous function arguments of nm_modem_act_stage2_config() act as if the
function could fail or even postpone the action. It never did.

We cannot treat this generic. A caller needs to know whether nm_modem_act_stage2_config()
can postpone the action, and when it does, which signal is emitted upon completion. That
is, the caller needs to know how to proceed after postponing.

In other words, since this function currently cannot fail or postpone
the stage, so must all callers already rely on that. At this point it makes
no sense to pretend that the function could be any different, if all callers
assume it is not. Simplify the API.