libnm: drop nm_dbus_new_connection() helper API
We don't need a wrapper around g_bus_get*(). Just use
it directly.

I guess in the past this had some use when we were using
a private socket too. Those days are gone. If we are going
to re-introduce private socket support, then we probably should
come up with a better solution.