libnm: drop unused NM_SETTING_PARAM_GENDATA_BACKED property flag
The idea was that properties that are implemented via GENDATA still
could have a GObject property. As such, they would be marked with
this flag.

Currently, gendata properties are only implemented by NMSettingEthtool,
and there are no GObject properties where this flag is used. While it
might make sense in theory or in the future, it is unused.

Drop it.