automation: fetch NetworkManager's gitlab repository and add refs from merge-requests
Fetch both from github and gitlab, and also fetch the references for the
merge-requests/pull-requests.

In practice is github just a mirror of gitlab, so seemingly it wouldn't
make sense to fetch from there. However:

  - by adding github as a remote, we can preferably fetch from there.
    I think that is faster than our gitlab at freedesktop.org. Thank
    you, Microsoft!

  - pull requests against github are indeed not found in gitlab.

We need to fetch all kinds of remote references, so that the referenced
commits are in the git repository. Also, we need to fetch them under
various name, so that these references are available to CI.

For example, when someone opens a merge-request from their fork on
gitlab/github, the commit is usually not not referenced by regular
branches on gitlab/github. Hence, we couldn't schedule CI for those
commit. Also fetch the special references for these.

Also, don't use `timeout` to fetch the repository.