libnm/tests: add test for checking types
If a "nm_${TYPE}_class_init()" function has a bug, then this
code only runs when the class gets instanciated. And for types
like NMDeviceBridge, that didn't happen for unit tests (so far).

Instanciate all glib types. In the future we may want to perform
additional checks on the types.